% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{as.cascade.data.frame}
\alias{as.cascade.data.frame}
\title{Create cascade object from data frame}
\usage{
\method{as.cascade}{data.frame}(dat, node_ids, node_names = NULL)
}
\arguments{
\item{dat}{\link{data.frame} with three columns containing the cascade 
infromation. The first column contains the node ids of the cascades, the 
second column the infection times of the corresponding nodes and the third
column contains an integer cascade identifier}

\item{node_ids}{Vector of integer ids for each node.}

\item{node_names}{Character vector of names for each node. Optional. Must be 
   of same length and sorting as node_ids. If not provided node_ids are used
   as names.

@import checkmate}
}
\description{
Create cascade object from data frame
}

