% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{as.cascade}
\alias{as.cascade}
\title{Create a cascade object from input data}
\usage{
as.cascade(dat, node_names = NULL)
}
\arguments{
\item{dat}{Cascades to be converted. See Details for supported classes.}

\item{node_names}{Character, numeric or factor vector of names for each node. 
Optional.}
}
\value{
An object of class \code{cascade}. This is a list containing three
    (named) elements: 
    \enumerate{
        \item \code{node_names} A character vector of node names.
        \item \code{cascade_nodes} A list with one character vector per
            cascade containing the node names in order of the events.
        \item \code{cascade_times} A list with one element per cascade 
            containing the event times for the nodes in \code{cascade_names}.
    }
}
\description{
A generic function to transform input data into a cascade object to be used 
in other \code{NetworkInference} function. The method invoked depends on the 
\link{class} of the first argument. See section Details for available methods.
}
\details{
The following methods are available:
\itemize{
    \item \link{as.cascade.data.frame}
    \item \link{as.cascade.matrix}
}
}

