% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{as.data.frame.cascade}
\alias{as.data.frame.cascade}
\title{Convert a cascade object to a data frame}
\usage{
\method{as.data.frame}{cascade}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Cascade object to convert.}

\item{row.names}{NULL or a character vector giving the row names for the data 
frame. Missing values are not allowed. (Not supported)}

\item{optional}{logical. If TRUE, setting row names and converting column 
names (to syntactic names: see make.names) is optional. (Not supported)}

\item{...}{Additional arguments passed to \code{\link{data.frame}}.}
}
\value{
A data frame with three columns. Containing (in order) 1) The names of 
    the nodes that experience an event in each cascade, 2) the event time of the
    corresponding node, 3) the cascade identifier.
}
\description{
Generates a data frame containing the cascade information in the cascade object.
Node information is lost
}

