% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{assert_cascade_consistency_}
\alias{assert_cascade_consistency_}
\title{Assert cascade consistency}
\usage{
assert_cascade_consistency_(cascade_nodes, cascade_times, node_names)
}
\arguments{
\item{cascade_nodes}{List of vectors of integer ids in order of infection.}

\item{cascade_times}{List of vectors of infection times corresponding to 
\code{cascade_nodes}.}

\item{node_names}{Full list of node names.}
}
\description{
Assert that the cascade information provided by the user is consistent.
}
