% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{assert_node_info_}
\alias{assert_node_info_}
\title{Assert that node information is consistent}
\usage{
assert_node_info_(node_ids, node_names)
}
\arguments{
\item{node_ids}{Integer node ids.}

\item{node_names}{Character node names.}
}
\description{
Used in all methods invoked by as.cascade to check if the node information 
(ids, times) passed into the funciton by the user are valid.
}

