% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cascades}
\alias{cascades}
\title{Example cascades}
\format{An object of class \code{cascade} containing 4 objects
\describe{
  \item{node_names}{Character node names}
  \item{cascade_nodes}{A list of integer vectors. Each containing the names of the
      nodes infected in this cascades in the order of infection}
  \item{cascade_times}{A list of numeric vectors. Each containing the infection
      times for the corresponding nodes in cascade_nodes}
}}
\source{
\url{https://github.com/snap-stanford/snap/blob/master/examples/netinf/example-cascades.txt}
}
\usage{
data(cascades)
}
\description{
An example dataset of 31 nodes and 54 cascades. From the original netinf 
implementation in SNAP.
}
\keyword{datasets}
