% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{count_dups_}
\alias{count_dups_}
\title{Count duplicates in a vector}
\usage{
count_dups_(x)
}
\arguments{
\item{x}{vector to count duplicates in}
}
\value{
An integer count of the number of duplicated values.
}
\description{
Count duplicates in a vector
}
