% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{count_possible_edges_}
\alias{count_possible_edges_}
\title{Count the number of possible edges given the data}
\usage{
count_possible_edges_(cascade_nodes, cascade_times)
}
\arguments{
\item{cascade_nodes}{List of integer vectors of node ids in order of event, 
one per cascade.}

\item{cascade_times}{List of numeric vectors of event times corresponding to
nodes in cascade_nodes (same order).}
}
\value{
Integer number of possible edges
}
\description{
Count the number of possible edges given the data
}
