% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edge_weight_}
\alias{edge_weight_}
\title{Calculate the edge weight between two nodes}
\usage{
edge_weight_(event_time_i, event_time_j, lambda, beta = 0.5,
  epsilon = 1e-11, tied = FALSE, model = 1L)
}
\description{
Calculate the edge weight between two nodes
}

