% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{is.cascade}
\alias{is.cascade}
\title{Is the object of class cascade?}
\usage{
is.cascade(object)
}
\arguments{
\item{object}{the object to be tested.}
}
\value{
\code{TRUE} if object is a cascade, \code{FALSE} otherwise.
}
\description{
Is the object of class cascade?
}
\examples{

data(cascades)
is.cascade(cascades)
# > TRUE
is.cascade(1) 
# > FALSE
}
