% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netinf.R
\name{is.diffnet}
\alias{is.diffnet}
\title{Is the object of class diffnet}
\usage{
is.diffnet(object)
}
\arguments{
\item{object}{The object to be tested}
}
\value{
\code{TRUE} if object is a diffnet, \code{FALSE} otherwise.
}
\description{
Is the object of class diffnet
}
\examples{

data(cascades)
result <- netinf(cascades, n_edges = 6, lambda = 1)
is.diffnet(result)
}
