% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netinf.R
\name{netinf}
\alias{netinf}
\title{Infer latent diffusion network}
\usage{
netinf(node_ids, node_names, cascade_ids, cascade_times,
  trans_mod = "exponential", alpha = 1, n_iter = 5)
}
\arguments{
\item{node_ids}{An vector of integer node ids.}

\item{node_names}{A charcter vector of node names.}

\item{cascade_ids}{A list of integer vectors containing the node ids of
the cascade in order of infection.}

\item{cascade_times}{A list of numeric vectors each containing infection 
times for the correspoinding nodes in \code{cascade_ids}.}

\item{trans_mod}{character, indicating the choice of model: 
\code{"exponential"}, \code{"power"} (power law) or \code{"rayleigh"}.}

\item{alpha}{Numeric, alpha for transmission model.}
}
\description{
Infer the most likely latent diffusion model from a set of nodes and 
infection times
}

