% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netinf_just_R.R
\name{netinfR}
\alias{netinfR}
\title{Pure R netinf implementation}
\usage{
netinfR(cascade_data, n.edges, lambda, beta = 0.5, epsilon = 10^(-9))
}
\arguments{
\item{cascade_data, }{a matrix with three columns (cascade id, then vertex id, 
then infection time)}
}
\description{
Slow. For testing only.
}

