% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimal_spanning_tree_}
\alias{optimal_spanning_tree_}
\title{Calculate the optimal spanning tree for a cascade}
\usage{
optimal_spanning_tree_(this_cascade_ids, this_cascade_times, lambda, beta,
  epsilon)
}
\description{
Calculate the optimal spanning tree for a cascade
}

