% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{order_cascade_}
\alias{order_cascade_}
\title{Sort cascades by event time}
\usage{
order_cascade_(cascades)
}
\arguments{
\item{cascades, }{object of class cascade}
}
\value{
An object of class cascade with each cascade (ids and times) ordered
    by event time
}
\description{
Sort cascades by event time
}

