% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{PLOT_THEME_}
\alias{PLOT_THEME_}
\title{Common plotting elements}
\usage{
PLOT_THEME_(mode = NULL)
}
\arguments{
\item{mode}{What elements to return.}
}
\value{
A ggplot object that can be added to a ggplot plot
}
\description{
Plotting layout for NetworkInference package.
}
