% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim_validation}
\alias{sim_validation}
\title{Larger simulated validation network.}
\format{An object of class \code{data.frame} with 4 columns, containing:
\describe{
  \item{origin_node}{Origin of diffusion edge.}
  \item{destination_node}{Destination node of diffusion edge.}
  \item{improvement}{Improvement in score for the edge}
  \item{p-value}{p-value for vuong test}
}}
\source{
See code below.
}
\usage{
data(sim_validation)
}
\description{
A network from simulated data. For testing purposes.
}
\keyword{datasets}
