% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{simulate_cascades_}
\alias{simulate_cascades_}
\title{Simulate a set of cascades}
\usage{
simulate_cascades_(n_cascades)
}
\arguments{
\item{n_cascades}{Number of cascades to generate}
}
\description{
Simulate a set of cascades
}

