% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cascade}
\alias{summary.cascade}
\title{Summarize cascade object}
\usage{
\method{summary}{cascade}(cascades)
}
\arguments{
\item{cascades}{Object of class cascade to be plottet}
}
\value{
Prints cascade summary information to the screen and returns summary
    stats in a list
}
\description{
Summarize cascade object
}

