% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{validation}
\alias{validation}
\title{Validation output from netinf source.}
\format{An object of class \code{data.frame} with 6 columns, containing:
\describe{
  \item{origin_node}{Origin of diffusion edge.}
  \item{destination_node}{Destination node of diffusion edge.}
  \item{volume}{??}
  \item{marginal_gain}{Marginal gain from edge.}
  \item{median_time_difference}{Median time between events in origin and 
      destination}
  \item{mean_time_difference}{Mean time between events in origin and 
      destination}
}}
\source{
Output from netinf example program (\url{https://github.com/snap-stanford/snap/tree/master/examples/netinf}).
}
\usage{
data(validation)
}
\description{
Contains output from original netinf C++ implementation, executed on 
\code{\link{cascades}}. For testing purposes.
}
\keyword{datasets}
