% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{as.cascade.data.frame}
\alias{as.cascade.data.frame}
\title{Create cascade object from data frame}
\usage{
\method{as.cascade}{data.frame}(dat, node_names = NULL)
}
\arguments{
\item{dat}{\link{data.frame} with three columns containing the cascade 
information. The first column contains the node names of the cascades, the 
second column the event times of the corresponding nodes and the third
column contains a cascade identifier (can be \code{integer}, 
\code{numeric}, \code{character} of \code{factor})}

\item{node_names}{Character, factor or numeric vector of names for each node. 
Optional. If not provided, node names are inferred from the cascade data.
Note that in this case nodes that are not involved in any cascade (isolates)
will be dropped (not recommended).}
}
\value{
An object of class \code{cascade}.
}
\description{
Create cascade object from data frame
}

