% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{assert_cascade_consistency_}
\alias{assert_cascade_consistency_}
\title{Assert cascade consistency}
\usage{
assert_cascade_consistency_(ids, times)
}
\arguments{
\item{ids}{List of vectors of integer ids in order of infection.}

\item{times}{List of vectors of infection times corresponding to \code{ids}.}
}
\description{
Assert that the cascade information provided by the user is consistent.
}

