% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netinf.R
\name{count_possible_edges}
\alias{count_possible_edges}
\title{Count the number of possible edges in the dataset}
\usage{
count_possible_edges(cascades)
}
\arguments{
\item{cascades}{Object of class cascade containing the data.}
}
\value{
An integer count.
}
\description{
Across all cascades, count the edges that are possible. An edge from node
\code{u} to node \code{v}
is only possible if in at least one cascade \code{u} experienced an event 
before \code{v}.
}
\examples{
data(cascades)
count_possible_edges(cascades)

}
