% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netinf.R
\name{is.diffnet}
\alias{is.diffnet}
\title{Is the object of class diffnet?}
\usage{
is.diffnet(object)
}
\arguments{
\item{object}{the object to be tested.}
}
\value{
\code{TRUE} if object is a diffnet, \code{FALSE} otherwise.
}
\description{
Tests if an object is of class diffnet. The class diffnet is appended to the 
object returned by \code{\link{netinf}} for dispatch of appropriate plotting 
methods.
}
\examples{

data(cascades)
result <- netinf(cascades, n_edges = 6, params = 1)
is.diffnet(result)
}
