% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netinf.R
\name{netinf}
\alias{netinf}
\title{Infer latent diffusion network.}
\usage{
netinf(cascades, trans_mod = "exponential", alpha = 1, n_iter = 5,
  verbose = TRUE)
}
\arguments{
\item{cascades}{An object of class cascade containing node and cascade 
information. See \link{as.cascade} for details.}

\item{trans_mod}{character, indicating the choice of model: 
\code{"exponential"}, \code{"power"} (power law) or \code{"rayleigh"}.}

\item{alpha}{Numeric, alpha for transmission model.}

\item{n_iter}{Numeric, number of iterations for optimization.}

\item{verbose}{logical, should additional output be printed.}
}
\value{
Returns the inferred diffusion network as an object of class 
 \link[base]{matrix} of dimension \code{k x 2} where k is the number of edges.
}
\description{
Infer the most likely latent diffusion model from a set of nodes and 
infection times
}
\details{
The algorithm is describe in detail in Gomez-Rodriguez et al. (2010). 
Additional information as well as the C++ source can be found on the 
netinf website (\url{http://snap.stanford.edu/netinf/}).'
}
\examples{
data(cascades)
out <- netinf(cascades, trans_mod = "exponential", alpha = 1, verbose = TRUE)
              
}
\references{
M. Gomez-Rodriguez, J. Leskovec, A. Krause. Inferring Networks of Diffusion 
and Influence.The 16th ACM SIGKDD Conference on Knowledge Discovery and 
Data Mining (KDD), 2010.
}

