% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netinf.R
\name{netinf}
\alias{netinf}
\title{Infer latent diffusion network.}
\usage{
netinf(node_ids, node_names, cascade_ids, cascade_times,
  trans_mod = "exponential", alpha = 1, n_iter = 5)
}
\arguments{
\item{node_ids}{A vector of integer node ids.}

\item{node_names}{A charcter vector of node names.}

\item{cascade_ids}{A list of integer vectors containing the node ids of
the cascade in order of infection.}

\item{cascade_times}{A list of numeric vectors each containing infection
times for the correspoinding nodes in \code{cascade_ids}.}

\item{trans_mod}{character, indicating the choice of model:
\code{"exponential"}, \code{"power"} (power law) or \code{"rayleigh"}.}

\item{alpha}{Numeric, alpha for transmission model.}
}
\value{
Returns the inferred diffusion network as an object of class
 \link[base]{matrix} of dimension \code{k x 2} where k is the number of edges.

The algorithm is describe in detail in Gomez-Rodriguez et al. (2010).
Additional information as well as the C++ source can be found on the
netinf website \url{http://snap.stanford.edu/netinf/}.
}
\description{
Infer the most likely latent diffusion model from a set of nodes and
infection times
}
\examples{
node_ids <- c(1:20)
node_names <- as.character(c(1:20))
cascade_ids <- list(c(2, 5, 12, 7),
                    c(3, 18, 6, 4, 7, 2),
                    c(4, 1, 7, 3, 12, 14, 2))
cascade_times <- list(c(34.5, 23, 5.67, 0),
                      c(23,  20, 14.2, 10, 8.5, 0),
                      c(40.1, 34.2, 26.9, 12.5, 10.5, 5, 0))
out <- netinf(node_ids = node_ids, node_names = node_names,
              cascade_ids = cascade_ids, cascade_times = cascade_times,
              trans_mod = "exponential", alpha = 1)
}
\references{
M. Gomez-Rodriguez, J. Leskovec, A. Krause. Inferring Networks of Diffusion
and Influence.The 16th ACM SIGKDD Conference on Knowledge Discovery and
Data Mining (KDD), 2010.
}

