% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/netinf.R
\name{netinf}
\alias{netinf}
\title{Infer latent diffusion network.}
\usage{
netinf(node_ids, node_names, cascade_ids, cascade_times,
  trans_mod = "exponential", alpha = 1, n_iter = 5, verbose = TRUE)
}
\arguments{
\item{node_ids}{A vector of integer node ids.}

\item{node_names}{A charcter vector of node names.}

\item{cascade_ids}{A list of integer vectors containing the node ids of
the cascade in order of infection.}

\item{cascade_times}{A list of numeric vectors each containing infection
times for the correspoinding nodes in \code{cascade_ids}.}

\item{trans_mod}{character, indicating the choice of model:
\code{"exponential"}, \code{"power"} (power law) or \code{"rayleigh"}.}

\item{alpha}{Numeric, alpha for transmission model.}

\item{verbose}{logical, should additional output be printed.}
}
\value{
Returns the inferred diffusion network as an object of class
 \link[base]{matrix} of dimension \code{k x 2} where k is the number of edges.
}
\description{
Infer the most likely latent diffusion model from a set of nodes and
infection times
}
\details{
The algorithm is describe in detail in Gomez-Rodriguez et al. (2010).
Additional information as well as the C++ source can be found on the
netinf website (\url{http://snap.stanford.edu/netinf/}).'
}
\examples{
data(example_cascades)
out <- netinf(node_ids = example_cascades$node_ids,
              node_names = example_cascades$node_names,
              cascade_ids = example_cascades$cascade_ids,
              cascade_times = example_cascades$cascade_times,
              trans_mod = "exponential", alpha = 1, verbose = TRUE)
}
\references{
M. Gomez-Rodriguez, J. Leskovec, A. Krause. Inferring Networks of Diffusion
and Influence.The 16th ACM SIGKDD Conference on Knowledge Discovery and
Data Mining (KDD), 2010.
}

