% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{netinf_}
\alias{netinf_}
\title{Run the netinf algorithm on a set of nodes and cascades}
\usage{
netinf_(node_ids, node_names, cascade_ids, cascade_times, model = 0L,
  alpha = 1, n_iter = 5L, verbose = TRUE)
}
\arguments{
\item{node_ids}{An integer vector of integer node ids.}

\item{node_names}{A charcter vector of node names.}

\item{cascade_ids}{A list of integer vectors containing the node ids of
the cascade in order of infection.}

\item{cascade_times}{A list of numeric vectors each containing infection
times for the correspoinding nodes in \code{cascade_ids}.}

\item{model}{integer indicating the choice of model: 0: exponential,
1: power law, 2: rayleigh.}

\item{alpha}{Numeric, alpha for transmission model.}

\item{verbose}{boolean, should additional information be printed.}
}
\description{
Run the netinf algorithm on a set of nodes and cascades
}

