% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{netinf_}
\alias{netinf_}
\title{Run the netinf algorithm on a set of nodes and cascades}
\usage{
netinf_(node_ids, cascade_nodes, cascade_times, n_edges, model, lambda)
}
\arguments{
\item{node_ids}{An integer vector of integer node ids.}

\item{cascade_nodes}{A list of integer vectors containing the node ids of
the cascade in order of infection.}

\item{cascade_times}{A list of numeric vectors each containing infection 
times for the corresponding nodes in \code{cascade_ids}.}

\item{n_edges}{Numeric, number of edges to infer.}

\item{model}{integer indicating the choice of model: 1: exponential, 
2: power law, 3: rayleigh (only exponential implemented).}

\item{lambda}{Numeric, rate parameter for exponential transmission model.}
}
\value{
List containing one vector per edge.
}
\description{
Run the netinf algorithm on a set of nodes and cascades
}

