% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.cascade}
\alias{plot.cascade}
\title{Plot a cascade object}
\usage{
\method{plot}{cascade}(cascades, label_nodes = TRUE, selection = NULL,
  plot_elements = NULL, density_ = TRUE)
}
\arguments{
\item{cascades}{Object of class cascade to be plottet}

\item{label_nodes}{Logical, indicating if should the nodes in each cascade be 
labeled. If the cascades are very dense setting this to \code{FALSE} is
recommended.}

\item{selection}{A vector of cascade ids to plot.}

\item{plot_elements}{Addtional ggplot plotting elements to be appended to the
plot (e.g. axis labels etc.).}

\item{density_}{draw density lines for each cascade.}
}
\value{
A ggplot plot object
}
\description{
Plot a cascade object
}

