% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{policies}
\alias{policies}
\title{US States Policy Adaption}
\format{The data is in a matrix format. Rows correspond to states (see rownames) columns
    to policies. Cell entries indicate the year a state adopted a policy. \code{NA} 
    entries indicate states not having adopted a policy at all.}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/A1GIMB}
}
\usage{
data(policies)
}
\description{
Adoption dates (years) for 187 policies across 50 US states. Compiled by
Desmarais et al. (2015).
}
\references{
Desmarais, B. A., Harden, J. J., & Boehmke, F. J. (2015). 
    Persistent Policy Pathways: Inferring Diffusion Networks in the American States. 
    American Political Science Review, 109(02), 392-406.
}
\keyword{datasets}
