% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascade.R
\name{simulate_cascades}
\alias{simulate_cascades}
\title{Simulate a set of cascades}
\usage{
simulate_cascades(n_cascades, n_nodes, id_class = "character")
}
\arguments{
\item{n_cascades}{Number of cascades to generate.}

\item{n_nodes}{Number of nodes in the system.}

\item{id_class}{One of \code{c("character", "factor", "numeric")}. What class
should the cascade_id indicator be.}
}
\value{
A data frame containing (in order of columns) infected node ids, 
    infection times and cascade identifiers for 20 distinct nodes.
}
\description{
Simulate random cascades, for testing and demonstration purposes. No actual 
diffusion model is underlying these cascades.
}
\examples{

df <- simulate_cascades(10, n_nodes = 20)
head(df)

}

