% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cascades.R
\name{simulate_cascades}
\alias{simulate_cascades}
\title{Simulate cascades from a diffusion network}
\usage{
simulate_cascades(diffnet, nsim = 1, seed = NULL, max_time = Inf, lambda,
  beta, epsilon, model)
}
\arguments{
\item{diffnet}{Object of class \code{diffnet}.}

\item{nsim}{Number of cascades to simulate.}

\item{seed}{Seed for random number generator.}

\item{max_time}{Numeric, the maximum time after which observations are 
censored}

\item{lambda}{Parameter for diffusion time distribution.}

\item{beta}{Weight for in-network diffusion}

\item{epsilon}{Weight for out of network diffusion}

\item{model}{Diffusion model to use. One of \code{c("exponential", 
"rayleigh")}.}
}
\value{
A data frame containing (in order of columns) node ids, 
    event time and cascade identifier.
}
\description{
Simulate random cascades, for testing and demonstration purposes. No actual 
diffusion model is underlying these cascades. Requires the \code{igraph} 
package to be loaded.
}
\examples{

data(cascades) 
out <- netinf(cascades, trans_mod = "exponential", n_edges = 5, lambda = 1)
simulated_cascades <- simulate_cascades(out, nsim = 10, lambda = 1, 
                                        beta = 0.5, epsilon = 10^-9, 
                                        model = "exponential")
}

