% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cascade}
\alias{summary.cascade}
\title{Summarize cascade object}
\usage{
\method{summary}{cascade}(object, ...)
}
\arguments{
\item{object}{Object of class cascade to be summarized.}

\item{...}{Additional arguments passed to summary}
}
\value{
Prints cascade summary information to the screen and returns summary
    stats in a list
}
\description{
Summarize cascade object
}
\examples{

data(cascades)
summary(cascades)

}

